Question 5 et question 3 (manque plateformes)

+++ Tp/api/Index/deleteJeu.php
@@ -2,6 +2,16 @@
 require_once __DIR__."/../../Php/config.php";
 
 try{
+    $stmt= $conn->prepare("DELETE FROM `jeux_plateformes` WHERE `id_jeux`=:id");
+    $stmt->bindValue(":id", $id);
+    $stmt->execute();
+
+    if(!$stmt->rowCount()){
+        http_response_code(400);
+        echo "L'identifiant du jeu est invalide!";
+        exit;
+    }
+
     $stmt= $conn->prepare("DELETE FROM `jeux` WHERE `id`=:id");
     $stmt->bindValue(":id", $id);
     $stmt->execute();
@@ -11,6 +21,8 @@ try{
         echo "L'identifiant du jeu est invalide!";
         exit;
     }
+    
+    
 
     $reponse = ["response"=>"OK"];
     header('Content-Type: application/json; charset=utf-8');
